requirements.txt
streamlit>=1.36.0
transformers>=4.44.0
torch>=2.2.0
PyMuPDF>=1.24.0
pillow>=10.0.0
numpy>=1.26.0
